Move tests and apply suggestions
